Removed control_view test case and renamed users.py to views.py
